Moved imports around and fixed folder architecture
